* Added Sample output wrapper IOWrapper/OutputWrapper/SampleOutputWrapper.h. * Added extensive comments to IOWrapper/Output3DWrapper.h. * Added support for multiple 3DOutputWrapper simulataneously. * Added options "quiet=1" and "sampleoutput=1". * Did some minor code cleaning.

+++ src/IOWrapper/OpenCV/ImageDisplay_OpenCV.cpp
@@ -141,37 +141,37 @@ void displayImage(const char* windowName, const MinimalImageB16* img, bool autoS
 void displayImageStitch(const char* windowName, const std::vector<MinimalImageB*> images, int cc, int rc)
 {
 	std::vector<cv::Mat*> imagesCV;
-	for(int i=0; i < images.size();i++)
+    for(size_t i=0; i < images.size();i++)
 		imagesCV.push_back(new cv::Mat(images[i]->h, images[i]->w, CV_8U, images[i]->data));
 	displayImageStitch(windowName, imagesCV, cc, rc);
-	for(int i=0; i < images.size();i++)
+    for(size_t i=0; i < images.size();i++)
 		delete imagesCV[i];
 }
 void displayImageStitch(const char* windowName, const std::vector<MinimalImageB3*> images, int cc, int rc)
 {
 	std::vector<cv::Mat*> imagesCV;
-	for(int i=0; i < images.size();i++)
+    for(size_t i=0; i < images.size();i++)
 		imagesCV.push_back(new cv::Mat(images[i]->h, images[i]->w, CV_8UC3, images[i]->data));
 	displayImageStitch(windowName, imagesCV, cc, rc);
-	for(int i=0; i < images.size();i++)
+    for(size_t i=0; i < images.size();i++)
 		delete imagesCV[i];
 }
 void displayImageStitch(const char* windowName, const std::vector<MinimalImageF*> images, int cc, int rc)
 {
 	std::vector<cv::Mat*> imagesCV;
-	for(int i=0; i < images.size();i++)
+    for(size_t i=0; i < images.size();i++)
 		imagesCV.push_back(new cv::Mat(images[i]->h, images[i]->w, CV_32F, images[i]->data));
 	displayImageStitch(windowName, imagesCV, cc, rc);
-	for(int i=0; i < images.size();i++)
+    for(size_t i=0; i < images.size();i++)
 		delete imagesCV[i];
 }
 void displayImageStitch(const char* windowName, const std::vector<MinimalImageF3*> images, int cc, int rc)
 {
 	std::vector<cv::Mat*> imagesCV;
-	for(int i=0; i < images.size();i++)
+    for(size_t i=0; i < images.size();i++)
 		imagesCV.push_back(new cv::Mat(images[i]->h, images[i]->w, CV_32FC3, images[i]->data));
 	displayImageStitch(windowName, imagesCV, cc, rc);
-	for(int i=0; i < images.size();i++)
+    for(size_t i=0; i < images.size();i++)
 		delete imagesCV[i];
 }
